% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_perm_palindrome.R
\name{is_perm_palindrome}
\alias{is_perm_palindrome}
\title{is_perm_palindrome}
\usage{
is_perm_palindrome(string)
}
\description{
it checks given string is palindrome or not by checking 1 element
from back and 1 element from beginning. If they are
same, function writes "it is not a palindrome. If they are
not same, function checks every element until string is  done.
If string is palindrome, function writes it is a palindrome
}
