% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_of_the_others.R
\name{product_of_the_others}
\alias{product_of_the_others}
\title{product_of_the_others}
\usage{
product_of_the_others(a)
}
\description{
it multiplies all of the elements in vector except the
element that you want and writes the multiplication result
}
